close #16 dont remove package in local

+++ Install.vala
@@ -244,7 +244,7 @@ public void install() throws Error {
 				else
 					install_suprapackage(i.output);
 			}
-			if (!config.is_cached) {
+			if (!config.is_cached && i.repo != "Local") {
 				FileUtils.unlink(i.output);
 			}
 		}
